+++ changelog.md
@@ -4,6 +4,7 @@
 
 * Prevent websocket reconnect loop when the authentication fails.
 * Added gunicorn worker test to the CI pipeline.
+* Made it possible to re-order fields in a view.
 
 ## Released (2021-03-01)